refactor: updates core folder to lib
