+++ docs/src/content/docs/recipes/sentry-setup.mdx
@@ -170,7 +170,7 @@ The starter kit did not come with Sentry pre-configured, but it's very easy to s
    Then, initialize Sentry and configure it with navigation in your `src/app/_layout.tsx` file:
 
    ```tsx title='src/app/_layout.tsx'
-   import { initSentry, useSentryNavigationConfig } from '@/core/sentry';
+   import { initSentry, useSentryNavigationConfig } from '@/lib/sentry';
    import * as Sentry from '@sentry/react-native';
 
    initSentry();
