+++ src/lib/i18n/react-i18next.d.ts
