+++ src/lib/i18n/resources.ts
